pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synchrob"
version = "0.1.0"
description = "SynchroB: Synchronized product behavior analysis and API generation system"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "SynchroB Team"}
]
keywords = ["api", "product-analysis", "code-generation", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "firecrawl-py>=0.0.1",
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "google-genai>=1.65.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "ruff>=0.1.0",
]
step3 = [
    "torch>=2.0.0",
    "torch-geometric>=2.3.0",
    "networkx>=3.1",
    "tree-sitter>=0.20.0",
    "tree-sitter-python>=0.20.0",
    "libcst>=1.1.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
]

[project.urls]
Homepage = "https://github.com/your-org/synchrob"
Documentation = "https://github.com/your-org/synchrob/blob/main/README.md"
Repository = "https://github.com/your-org/synchrob.git"
Issues = "https://github.com/your-org/synchrob/issues"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.isort]
known-first-party = ["src"]
